+++ CHANGELOG.rst
@@ -17,6 +17,17 @@ Unreleased
 ----------
 * nothing unreleased
 
+[8.14.0] - 2026-09-16
+----------------------
+* feat: Add ``ActivationEmailEnterpriseContextEnricher`` pipeline step for the
+  ``AccountActivationEmailContextGenerated`` filter (ENT-11816)
+
+[8.13.0] - 2026-09-14
+----------------------
+* Changed ``SupportEnterpriseEnrollmentDataInjector`` to accept/return the full enrollments list
+  and attach enterprise course enrollment data to each matching enrollment dict in place, instead
+  of returning a separate dict keyed by course_id that the platform had to zip back together itself
+
 [8.12.0] - 2026-09-08
 ----------------------
 * feat: add SupportEnterpriseEnrollmentDataInjector pipeline step (ENT-11574)